+++ blender/plane.js
@@ -19,7 +19,7 @@ export const uvs = new Float32Array([
     1, 0,
 ]);
 
-export const indices = new Uint16Array([
+export const indices = new Uint32Array([
     0, 1, 2,
     2, 1, 3
 ]);
